change reactive.runner_manager.reconcile signature
