+++ src/github_runner_manager/reactive/runner.py
@@ -7,12 +7,6 @@ import logging
 import os
 import sys
 
-from pydantic import BaseModel
-
-from github_runner_manager.manager.cloud_runner_manager import GitHubRunnerConfig
-from github_runner_manager.manager.runner_manager import RunnerManagerConfig
-from github_runner_manager.openstack_cloud.openstack_runner_manager import OpenStackCloudConfig, \
-    OpenStackServerConfig
 from github_runner_manager.reactive.consumer import consume, QueueConfig
 from github_runner_manager.reactive.runner_manager import MQ_URI_ENV_VAR, QUEUE_NAME_ENV_VAR
 
@@ -25,24 +19,6 @@ def setup_root_logging() -> None:
         level=logging.DEBUG,
         format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
     )
-
-
-class ReactiveRunnerConfig(BaseModel):
-    """The configuration for the reactive runner to spawn.
-
-    Attributes:
-        queue: The queue configuration.
-        runner_manager: The runner manager configuration.
-        runner: The GitHub runner configuration.
-        openstack_cloud: The OpenStack cloud configuration.
-        openstack_server: The OpenStack server configuration.
-    """
-
-    queue: QueueConfig
-    runner_manager: RunnerManagerConfig
-    runner: GitHubRunnerConfig
-    openstack_cloud: OpenStackCloudConfig
-    openstack_server: OpenStackServerConfig | None = None
 
 
 def main() -> None:
